break out classes for different types of annotators into separate submodules
